Refactor directory structure for better organization
1. Renamed data_sources to sources:
   - More concise name (7 characters vs 12)
   - Maintains clarity about purpose
   - Better fits with existing naming pattern (analysis, pipeline)
   - Creates a more consistent directory structure

2. Removed unused storage directory:
   - Contained only __pycache__ files
   - No actual Python source files
   - Functionality moved to sources/clients

3. Updated all imports throughout the codebase:
   - Modified imports in processor.py
   - Updated imports in health_pipeline.py
   - Fixed internal imports within sources directory

This change improves code organization while maintaining all functionality.
