+++ src/sources/base.py
@@ -14,7 +14,7 @@ import requests
 from src.utils.date_utils import DateFormat, DateUtils
 from src.utils.logging_utils import HealthLogger
 from src.utils.progress_indicators import ProgressIndicator
-from src.data_sources.token_manager import TokenManager
+from src.sources.token_manager import TokenManager
 
 
 class DataSource(ABC):
