+++ src/sources/clients/oura.py
@@ -11,7 +11,7 @@ from src.utils.date_utils import DateFormat, DateUtils
 from src.utils.file_utils import save_json_to_file
 from src.utils.logging_utils import HealthLogger
 
-from src.data_sources.token_manager import TokenManager
+from src.sources.token_manager import TokenManager
 
 
 class OuraError(Exception):
