+++ src/sources/clients/whoop.py
@@ -11,7 +11,7 @@ from urllib.parse import parse_qs, urlencode, urlparse
 
 import requests
 
-from src.data_sources.base import APIClient, APIClientError, OAuthCallbackHandler
+from src.sources.base import APIClient, APIClientError, OAuthCallbackHandler
 from src.utils.date_utils import DateFormat, DateUtils
 from src.utils.file_utils import save_json_to_file
 from src.utils.progress_indicators import ProgressIndicator
